+++ dao/dao-base/src/main/java/com/wpx/model/system/applicationtopic/pojo/cms/ApplicationTopicCmsAddDTO.java
@@ -22,7 +22,7 @@ public class ApplicationTopicCmsAddDTO implements Serializable {
 
     @ApiModelProperty(value = "应用ID")
     @NotNull(message = "应用ID不能为空")
-    private Integer applicationId;
+    private Long applicationId;
 
     @ApiModelProperty(value = "主题名称")
     @NotNull(message = "主题名称不能为空")
