+++ dao/dao-base/src/main/java/com/wpx/model/system/managerlog/ManagerLog.java
@@ -28,10 +28,10 @@ public class ManagerLog implements Serializable {
 
     @ApiModelProperty(value = "管理员操作日志ID")
     @TableId(value = "manager_log_id", type = IdType.AUTO)
-    private Integer managerLogId;
+    private Long managerLogId;
 
     @ApiModelProperty(value = "管理员ID")
-    private Integer managerId;
+    private Long managerId;
 
     @ApiModelProperty(value = "请求方法类型")
     private String method;
